Improved patron view of wait times
If a patron is in line the will not see the overall wait time, only
their wait time and the number of groups ahead of them in line. Also
fixed a bug in the accordian buttons, as well as fixed the NaN values on
waits with no tables of that size.